+++ package.json
@@ -9,6 +9,7 @@
   },
   "devDependencies": {
     "eslint": "^7.32.0",
+    "jscpd": "^3.3.26",
     "jsdoc": "^3.6.7",
     "standard": "^16.0.3",
     "webpack": "^5.52.1",
@@ -17,6 +18,7 @@
   "scripts": {
     "test": "eslint js/*.js",
     "jsdoc": "jsdoc src -d jsdoc/",
+    "jscpd": "jscpd ./src/**.js",
     "standard": "standard ./src/**.js",
     "standard:fix": "standard --fix ./js/bootstrap-validator.js",
     "super-linter": "docker run -e RUN_LOCAL=true -v `pwd`:/tmp/lint github/super-linter"
